clk: sunxi: display: support sun4i-a10-mali400-clk
This is the driver to be used for clock and reset for the Mali GPU on
the Allwinner A20.
It is fairly similar to the display clock register but has a different
number of input sources, so just use the same driver implementation with
different offset/width parameters.

Tested on Allwinner A20 (Cubieboard2) with mesa-lima.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>